requirements.txt
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
flask==2.0.1